Extract shared sim config logic and use tuned defaults
