+++ scripts/simulate_season_avg.py
@@ -57,7 +57,6 @@ from collections import Counter
 from datetime import date
 from pathlib import Path
 import argparse
-import csv
 import pickle
 import random
 import sys
@@ -90,10 +89,10 @@ from logic.simulation import (
     TeamState,
     generate_boxscore,
 )
-from logic.playbalance_config import PlayBalanceConfig
 from utils.lineup_loader import build_default_game_state
 from utils.path_utils import get_base_dir
 from utils.team_loader import load_teams
+from logic.sim_config import load_tuned_playbalance_config
 
 # Access ``save_stats`` for optional monkeypatching when running the script.
 import logic.simulation as sim
@@ -180,32 +179,6 @@ def _simulate_game_star(args: tuple[str, str, int]) -> Counter[str]:
     return _simulate_game(*args)
 
 
-def apply_league_benchmarks(
-    cfg: PlayBalanceConfig, benchmarks: dict[str, float]
-) -> None:
-    """Configure ``cfg`` using league-wide benchmark rates.
-
-    Parameters
-    ----------
-    cfg:
-        Play balance configuration instance to modify.
-    benchmarks:
-        Mapping of metric keys to numeric values loaded from
-        ``mlb_league_benchmarks_2025_filled.csv``.
-    """
-
-    hr_rate = cfg.hitHRProb / 100
-    # The raw league BABIP underestimates hits in our simplified physics model.
-    # Apply a modest boost to the baseline hit probability to better match
-    # observed MLB averages.
-    cfg.hitProbBase = benchmarks["babip"] / (1 - hr_rate) * 1.25
-    cfg.ballInPlayPitchPct = int(
-        round(benchmarks["pitches_put_in_play_pct"] * 100)
-    )
-    pitches_per_pa = benchmarks["pitches_per_pa"]
-    cfg.swingProbScale = round(4.0 / pitches_per_pa, 2) if pitches_per_pa else 1.0
-
-
 def simulate_season_average(
     use_tqdm: bool = True, ball_in_play_outs: int = 0, seed: int | None = None
 ) -> None:
@@ -233,54 +206,7 @@ def simulate_season_average(
             pickle.dump(schedule, fh)
     base_states = {tid: build_default_game_state(tid) for tid in teams}
 
-    cfg = PlayBalanceConfig.from_file(get_base_dir() / "logic" / "PBINI.txt")
-    cfg.ballInPlayOuts = ball_in_play_outs
-
-    csv_path = (
-        get_base_dir()
-        / "data"
-        / "MLB_avg"
-        / "mlb_avg_boxscore_2020_2024_both_teams.csv"
-    )
-    with csv_path.open(newline="") as f:
-        row = next(csv.DictReader(f))
-    hits = float(row["Hits"])
-    singles = (
-        hits
-        - float(row["Doubles"])
-        - float(row["Triples"])
-        - float(row["HomeRuns"])
-    )
-    cfg.hit1BProb = int(round(singles / hits * 100))
-    cfg.hit2BProb = int(round(float(row["Doubles"]) / hits * 100))
-    cfg.hit3BProb = int(round(float(row["Triples"]) / hits * 100))
-    cfg.hitHRProb = max(
-        0,
-        100 - cfg.hit1BProb - cfg.hit2BProb - cfg.hit3BProb,
-    )
-    at_bats = float(row["AtBats"])
-    walks = float(row["Walks"])
-    hbp = float(row["HitByPitch"])
-    total_pitches = float(row["TotalPitchesThrown"])
-    strikeouts = float(row["Strikeouts"])
-    homers = float(row["HomeRuns"])
-    plate_appearances = at_bats + walks + hbp
-    balls_in_play = at_bats - strikeouts - homers
-
-    bench_path = (
-        get_base_dir()
-        / "data"
-        / "MLB_avg"
-        / "mlb_league_benchmarks_2025_filled.csv"
-    )
-    with bench_path.open(newline="") as bf:
-        benchmarks = {
-            r["metric_key"]: float(r["value"])
-            for r in csv.DictReader(bf)
-        }
-
-    apply_league_benchmarks(cfg, benchmarks)
-    mlb_averages = {stat: float(val) for stat, val in row.items() if stat}
+    cfg, mlb_averages = load_tuned_playbalance_config(ball_in_play_outs)
 
     # Prepare list of (home, away, seed) tuples for multiprocessing
     rng = random.Random(seed)
