+++ tests/test_league_benchmarks.py
@@ -1,7 +1,7 @@
 import pytest
 
 from logic.playbalance_config import PlayBalanceConfig
-from scripts.simulate_season_avg import apply_league_benchmarks
+from logic.sim_config import apply_league_benchmarks
 
 
 def test_apply_league_benchmarks():
